Unify User and UserProfile and use username from cognito as identifier
