+++ src/models/User.ts
@@ -1,0 +1,11 @@
+export enum Groups {
+    ADMINS = 'ADMINS',
+    CURATORS = 'CURATORS',
+    RESEARCHERS = 'RESEARCHERS',
+}
+
+export interface User {
+    username: string;
+    email: string,
+    groups: Groups[],
+}
