+++ src/redux/app/selectors.ts
@@ -1,7 +1,7 @@
-import { UserProfile } from './slice';
+import {User} from "../../models/User.ts";
 import { RootState } from '../store';
 
-export const selectUserProfile: (state: RootState) => UserProfile | undefined = (
+export const selectUserProfile: (state: RootState) => User | undefined = (
     state,
 ) => state.app.userProfile;
