+++ src/redux/app/thunk.ts
@@ -1,16 +1,16 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
 import {signOut} from "aws-amplify/auth";
-import {UserProfile} from "./slice.ts";
+import {User} from "../../models/User.ts";
 import {client} from "../../utils/amplifyClient";
 
 
-export const getUserProfile = createAsyncThunk<UserProfile, void, { rejectValue: string }>(
+export const getUserProfile = createAsyncThunk<User, void, { rejectValue: string }>(
     'app/getUserProfile',
     async (_, {rejectWithValue}) => {
         try {
             const response = await client.queries.getUserProfile({});
             const data = response.data;
-            return (typeof data === 'string' ? JSON.parse(data) : data) as UserProfile;
+            return (typeof data === 'string' ? JSON.parse(data) : data) as User;
         } catch (error: unknown) {
             return rejectWithValue(error instanceof Error ? error.message : 'Failed to fetch user profile');
         }
