+++ src/redux/dataDownloads/tests/thunk.test.ts
@@ -2,8 +2,7 @@ import {beforeEach, describe, expect, it, vi} from 'vitest';
 import {getFilesFromS3Folder, handleDownload} from '../thunk.ts';
 import {getUrl, list} from 'aws-amplify/storage';
 import {client} from '../../../utils/amplifyClient.ts';
-import {Groups} from "../../manageUsers/slice.ts";
-import {UserProfile} from "../../app/slice.ts";
+import {User, Groups} from "../../../models/User.ts";
 
 // Mock Amplify Storage
 vi.mock('aws-amplify/storage', () => ({
@@ -84,7 +83,7 @@ describe('DataDownloads thunks', () => {
     });
 
     describe('handleDownload', () => {
-        const mockUser: UserProfile = {email: 'user@example.com', id: '123', groups: [Groups.RESEARCHERS]};
+        const mockUser: User = {email: 'user@example.com', username: '123', groups: [Groups.RESEARCHERS]};
         const payload = {s3FileKey: 'file.txt', user: mockUser};
 
         it('should fulfill on successful download process', async () => {
@@ -96,7 +95,7 @@ describe('DataDownloads thunks', () => {
             expect(result.meta.requestStatus).toBe('fulfilled');
 
             expect(client.models.DownloadEvent.create).toHaveBeenCalledWith(expect.objectContaining({
-                userId: mockUser.id,
+                userId: mockUser.username,
                 email: mockUser.email,
                 filename: payload.s3FileKey,
             }));
